Consolidate FakeDatasource into unittests.utils
